AOSP: build mke2fs for windows using android mingw library
Bug: 23686092
From AOSP commit: 83da0276c3ff0a1c33f9042798b319542e254060

+++ lib/blkid/list.h
@@ -146,7 +146,7 @@ _INLINE_ void list_splice(struct list_head *list, struct list_head *head)
  * @member:	the name of the list_struct within the struct.
  */
 #define list_entry(ptr, type, member) \
-	((type *)((char *)(ptr)-(unsigned long)(&((type *)0)->member)))
+	((type *)((char *)(ptr)-(unsigned long)(intptr_t)(&((type *)0)->member)))
 
 /**
  * list_for_each - iterate over elements in a list
